support pickle files saved prior to change
